Silicon Labs Matter Extension v2.8.0

+++ slc/stage_extension.py
@@ -22,18 +22,38 @@ import shutil
 import argparse
 
 # Directories to exclude from copying
-exclude_root_directories = [ "matter_extension/out", "matter_extension/slc/tools"]
+exclude_root_directories = [
+    'matter_extension/out',
+    'matter_extension/slc/tools'
+]
 
-exlude_submodules = ['simplicity_sdk', 'wifi_sdk', "matter_private", "third_party_hw_drivers_extension"]
+exclude_submodules = [
+    'simplicity_sdk',
+    'wifi_sdk',
+    'third_party_hw_drivers_extension'
+]
 
-#specific exlude from matter_sdk
-matter_sdk_exclude = ['matter_sdk/third_party','matter_sdk/zzz_generated','matter_sdk/examples/virtual-device-app','matter_sdk/src/controller/java/generated','matter_sdk/examples/chef/devices','matter_sdk/scripts/py_matter_idl/matter/idl/tests/outputs/large_all_clusters_app']
+# specific exlude from matter_sdk
+matter_sdk_exclude = [
+    'matter_sdk/third_party',
+    'matter_sdk/examples/virtual-device-app',
+    'matter_sdk/src/controller/java/generated',
+    'matter_sdk/examples/chef/devices',
+    'matter_sdk/scripts/py_matter_idl/matter/idl/tests/outputs/large_all_clusters_app'
+]
 
 # Specific includes for wiseconnect_wifi_bt_sdk
-wiseconnect_wifi_bt_sdk_includes = ['sapi']
+wiseconnect_wifi_bt_sdk_includes = [
+    'sapi'
+]
 
 # Specific includes for matter_support
-matter_support_includes = ['provision', 'tinycrypt', "si91x", "lwip"]
+matter_support_includes = [
+    'provision',
+    'tinycrypt',
+    'lwip'
+]
+
 
 def should_exclude(root, path):
     """
@@ -41,53 +61,48 @@ def should_exclude(root, path):
     """
     # Exclude hidden files and directories
     if path.startswith("."):
-        #print(f"Excluding hidden path: {path}")
+        # print(f"Excluding hidden path: {path}")
         return True
-    
+
     full_path = os.path.join(root, path)
-    
-    # Check for wiseconnect_wifi_bt_sdk specific includes
-    if "wiseconnect-wifi-bt-sdk/" in full_path:
-        if not any(include in full_path for include in wiseconnect_wifi_bt_sdk_includes):
-            #print(f"Excluding wiseconnect_wifi_bt_sdk path: {full_path}")
-            return True
 
     # Exclude submodules
     if "third_party/" in full_path:
-        if any(exclude in full_path for exclude in exlude_submodules):
-            #print(f"Excluding submodule path: {full_path}")
+        if any(exclude in full_path for exclude in exclude_submodules):
+            # print(f"Excluding submodule path: {full_path}")
             return True
-    
+
     # Check for matter_support/board-support/ specific includes
     if "matter_support/board-support/" in full_path or "matter_support/sdk-copies/" in full_path:
         if not any(include in full_path for include in matter_support_includes):
-            #print(f"Excluding matter_support path: {full_path}")
+            # print(f"Excluding matter_support path: {full_path}")
             return True
 
     # Check for matter_support specific includes
     if "matter_support/" in full_path and not ("matter_support/board-support" in full_path or "matter_support/sdk-copies" in full_path):
         if not any(include in full_path for include in matter_support_includes):
-            #print(f"Excluding matter_support_2 path: {full_path}")
+            # print(f"Excluding matter_support_2 path: {full_path}")
             return True
-    
+
     # Exclude matter_sdk directories
     if "matter_sdk/" in full_path:
         if any(exclude in full_path for exclude in matter_sdk_exclude):
-            #print(f"Excluding matter_sdk path: {full_path}")
+            # print(f"Excluding matter_sdk path: {full_path}")
             return True
-    
+
     # Exclude directories listed in exclude_directory
     for exclude in exclude_root_directories:
         if exclude in full_path:
-            #print(f"Excluding directory from exclude list: {full_path}")
+            # print(f"Excluding directory from exclude list: {full_path}")
             return True
 
     # Exclude directories if any part of the path matches exclude_directory
     if any(exclude in full_path.split(os.sep) for exclude in exclude_root_directories):
-        #print(f"Excluding directory from split path: {full_path}")
+        # print(f"Excluding directory from split path: {full_path}")
         return True
 
     return False
+
 
 def copy_directory(source_directory, target_location):
     """
@@ -104,32 +119,42 @@ def copy_directory(source_directory, target_location):
     for root, dirs, files in os.walk(source_directory):
         # Skip directories that should be excluded
         dirs[:] = [d for d in dirs if not should_exclude(root, d)]
-        
+
         # Copy files
         for file in files:
             # Skip hidden files
             if file.startswith("."):
                 continue
+            # Skip .slt files
+            if file.endswith(".slt"):
+                continue
             source_file = os.path.join(root, file)
+            if not os.path.exists(source_file):
+                print(f"Warning: Source file does not exist: {source_file}")
+                continue
             relative_path = os.path.relpath(root, source_directory)
             target_dir = os.path.join(target_location, relative_path)
             os.makedirs(target_dir, exist_ok=True)
             shutil.copy2(source_file, os.path.join(target_dir, file))
 
+
 def main():
     # Set up argument parser
-    parser = argparse.ArgumentParser(description="Copy directory to target location.")
-    parser.add_argument('target_location', type=str, help='The target location to copy the directory to.')
+    parser = argparse.ArgumentParser(
+        description="Copy directory to target location.")
+    parser.add_argument('target_location', type=str,
+                        help='The target location to copy the directory to.')
     args = parser.parse_args()
-    
+
     target_location = args.target_location
-    
+
     # Define the source directory as the current working directory
     source_directory = os.getcwd()
-    
+
     # Copy the root directory to the target location, excluding the specified directories
     copy_directory(source_directory, target_location)
     print(f"Copied files to {target_location}/matter_extension.")
+
 
 if __name__ == "__main__":
     main()
